Query artist page (marketplace) + Query product (admin)

+++ lib_artiexh-data-jpa/src/main/java/com/artiexh/data/jpa/entity/ProductEntity.java
@@ -119,6 +119,6 @@ public class ProductEntity {
 
 	@Builder.Default
 	@Column(name = "sold_quantity", nullable = false)
-	private Integer soldQuantity = 0;
+	private Long soldQuantity = 0L;
 
 }
